Add an Enforcement enum
This tracks enabled enforcements independently of the PipelineOptions
object.

Move utility methods to construct enforcements and bundle factories
based around enabled enforcements to the Enforcement enum.

Only apply Immutability Enforcement if Immutability Enforcement is
applicable to the producing PTransform.

Remove EncodabilityEnforcement, as its responsibilities are handled in
CloningBundleFactory.
